Explicit data types.

+++ analysis/model_creation_pixelwise_par.py
@@ -60,7 +60,10 @@ def conv_par(idxPrc, aryPngData, vecHrf, queOut):
     The pixel-wise design matrix is convolved with an HRF model.
     """
     # Array for function output (convolved pixel-wise time courses):
-    aryPixConv = np.zeros(np.shape(aryPngData))
+    aryPixConv = np.zeros(np.shape(aryPngData), dtype=np.float32)
+
+    # Explicity typing:
+    vecHrf = vecHrf.astype(np.float32)
 
     # Number of volumes:
     varNumVol = aryPngData.shape[1]
@@ -71,19 +74,20 @@ def conv_par(idxPrc, aryPngData, vecHrf, queOut):
     for idxPix in range(0, aryPngData.shape[0]):
 
         # Extract the current pixel time course:
-        vecDm = aryPngData[idxPix, :]
+        vecDm = aryPngData[idxPix, :].astype(np.float32)
 
         # In order to avoid an artefact at the end of the time series, we have
         # to concatenate an empty array to both the design matrix and the HRF
         # model before convolution.
-        vecZeros = np.zeros([100, 1]).flatten()
+        vecZeros = np.zeros([100, 1], dtype=np.float32).flatten()
         vecDm = np.concatenate((vecDm, vecZeros))
         vecHrf = np.concatenate((vecHrf, vecZeros))
 
         # Convolve design matrix with HRF model:
         aryPixConv[idxPix, :] = np.convolve(vecDm,
                                             vecHrf,
-                                            mode='full')[0:varNumVol]
+                                            mode='full'
+                                            )[0:varNumVol].astype(np.float32)
 
     # Create list containing the convolved pixel-wise timecourses, and the
     # process ID:
